Переносит отрисовку доски из основного файла в контроллер
